+++ src/ui/game/player/complex_visual_player.cpp
@@ -59,8 +59,6 @@ bool ComplexVisualPlayer::init(RemotePlayer* parent, bool isSecond) {
         .parent(this)
         .store(nameLabel);
 
-    nameLabel->setSingleBadgeMode(true);
-
     this->updateIcons(data.icons);
 
     if (!isSecond && settings.players.statusIcons) {