+++ changelog.md
@@ -1,8 +1,9 @@
 ## v1.6.1
 
-* Fix crash when pressing any key on Mac
+* Fix crash when pressing any key on Mac (menu crash not fixed yet :/)
 * Fix player list looking weird after leaving a room with a pinned level
 * Add verification for discord linking
+* Only show 1 badge near a player's name when in a level
 
 ## v1.6.0